Created player object, created team object, added authentication for action, created admin tests for player and teams

+++ src/main/java/com/pinewoods/score/tracker/dto/admin/PlayerDTO.java
@@ -1,4 +1,4 @@
 package com.pinewoods.score.tracker.dto.admin;
 
-public record PlayerDTO(String name, String team, int handicap) {
+public record PlayerDTO(String name, String team, double handicap) {
 }
